backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.116.1",
    "logging>=0.4.9.6",
    "mongomock>=4.3.0",
    "mongomock-motor>=0.0.36",
    "motor>=3.7.1",
    "passlib[bcrypt]>=1.7.4",
    "pytest>=8.4.1",
    "python-jose>=3.5.0",
]


[tool.pytest.ini_options]
pythonpath = "."
filterwarnings = [
    "ignore:'crypt' is deprecated and slated for removal in Python 3.13:DeprecationWarning"
]

